requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
Pillow==8.1.0
PyMySQL==1.0.2
python-dotenv==0.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3

pandas==1.2.4